Add dark and light mode functionality
Updated the Theme.tsx component and the getRandomColor function. I
wrapped the App.tsx component in a ThemeContext and updated all the
instances of getRandomColor function with a useEffect to immediately
reflect the toggled state. I also updated the which icon gets showed,
depending on the light or dark mode.